GenericGeometry and Physics now inherit from Subdomain
