+++ openpnm/core/__init__.py
@@ -3,3 +3,4 @@ from .Workspace import Workspace
 from .Project import Project
 from .ModelsMixin import ModelsMixin, ModelsDict
 from .Base import Base
+from .Subdomain import Subdomain
